to much help my brain...
summary finished the component GameObject system. (add remove get components).
changed most of the raw ptr to smart ptr.
The GameObjectManager has a vector of shard_ptr<GameObject> those ptr can get shard with GameObject::Get(id/name)
if a GameObject gets deleted, a flag on the gameobject will say that it was deleted and if there is still a ref to the gameobject that trys to modifi it in some way. an warning in the console will be displayed that shows that it was tryed to do a change on a deleted gameobject. same for components.
added all the default assets and made the project build agin with a scene where a cube is displayed with the texture fallback and a camera that can move around