need to do this
resolve size in precent in diffrent types like aviable (%a), percent w/h (%w/%h), percent major (%ma), percent(%mi). in the size calculations. unit enum shows whitch unit the current size has like pixels or what type of percent and than says how to interpret the the value. If pixels do nothing, percent use x as major and y as minor so that  the size is dynamic. w and percent are x and y and a as is it is